feat(test): add order test

+++ src/main/java/org/example/dto/order/CreateOrderByCustomerDto.java
@@ -12,7 +12,7 @@ public record CreateOrderByCustomerDto(
 
         @NotNull
         @Positive
-        Integer offeredPrice,
+        Double offeredPrice,
 
         @NotNull
         ZonedDateTime expectedDoneAt,
